Allow live translations to happen during retranslateAll
Summary:
We force the request to use interpreter when the request is of RequestKind::Profile, and when we need to dump out precise profile data.

Otherwise, "standard" requests should be allowed to create new code, including live translations when retranslateAll is pending/happening (but not when we serialize profile data).

Reviewed By: ottoni

Differential Revision: D13508793

fbshipit-source-id: 3ae24c6e45f9904b2776385782ae6683bacb0f43