Add thread locals whitelist test
Summary:
This test checks the set of thread-local variables in the HHVM namespace against a fixed list. Any diff that adds new thread locals will have to touch this whitelist. Locking these variables down is a requirement for moving to userland scheduling.

The test just uses nm to list symbol names and comm to compare it to the list. It takes about 100s to run on a dbg build and 10s on a dbgo or opt build. The whitelist also supports simple bash-style comments, so we can group related thread locals and put a comment on each block or inline with a single entry.

Any new locals are printed in the failure output, along with a message explaining what happened. Note that we must use a whitelist, not an exact match, because the opt build drops some symbols.

Reviewed By: mofarrell

Differential Revision: D13498871

fbshipit-source-id: 3d5d51ca217f40b5b3165468cdc06960b831144a