use request-injected-data for signal delivery
Summary:
The current implementation of signals assumes that the signal is always delivered to the thread where the script is running, which is not guaranteed by the OS.  It isn't a problem if there is only one thread running in script mode, but we may move the script out of the main thread.

This diff introduces a more reliable mechanism to deliver a few kinds of signals (through surprise flag) to requests.  Those signals are blocked at the process level, and we start a dedicated thread that checks pending signals and run callbacks registered for the signal (which generally set surprise flags to all request threads, but can also do other things).

This diff doesn't affect server mode.  The mechanism can later be used in server mode, e.g., a server can broadcast a signal to all request threads when it is low on memory, or is about to shutdown, a request can install a signal handler which drops cache, runs GC, does cleanups, etc.

Reviewed By: alexeyt, markw65

Differential Revision: D10417378

fbshipit-source-id: 32d2488147bd0f78935197ecf8b9facbf91c079e