Fix unset and ref-binding checks for property type enforcement
Summary:
The checks to see if we care about unsetting or ref-binding a property with a
type-hint used isCheckable(). However if isCheckable() returns true, the
property might still resolve to mixed (for example, if its a type
alias). Instead invent a new check, isMixedResolved(), which will resolve
type-aliases (and therefore might invoke the auto-loader) and use that for
unsetting or ref-binding checks.

Reviewed By: markw65

Differential Revision: D13523241

fbshipit-source-id: f3d3358280bd0510b0389582d7a4b80fd30af5ed