Break tracelets after ClsCns
Summary:
During region formation, the type of an Object will be exact; so the
class extracted from it will be exact, so the value of a ClsCns will
be known. But since we don't end up guarding on that type, during
translation, we'll only know that the value is uncounted-init.

In the new test case, this meant that form_region "knew" that $c was
of class C1, so it knew the exact value of the class constant, and
didn't break the tracelet.

During translation, this caused implDictKeysetIdx (which expects to
get specific data types for its arguments) to incorrectly emit a
ThrowInvalidArrayKey (and then in this specific case crash because the
tracelet ended unexpectedly - other cases just result in incorrect
code).

Reviewed By: ricklavoie

Differential Revision: D13144908

fbshipit-source-id: f3ab6870bdc559ffb107a5d051bf2ed968d33a37