New inference: array append
Summary:
The array append operation is heavily overloaded, which presents problems with constraint-based type inference when the array in question has an unsolved type variable type. However, it can be mimicked by treating
```
$x[] = e
```
as the following function
```
function array_append<Ta as CommonArrayType<Tv>,Tv>(Ta $array, Tv $value) : Ta
```
The main problem is that there is no `CommonArrayType` in the existing hierarchy. So instead, we create (on the fly) a union
```
Collection<Tv> | vec<Tv> | keyset<Tv> | array<Tv>
```
which covers all the existing cases of the `assign_array_append` operation. `Collection<_>` is the common supertype of *only* `Vector<Tv>`, `Set<Tv>`, and `Map<Pair<Tk,Tv>>`, and `array` handles `varray` too.

We might in future consider creating a new interface to capture this union.

Reviewed By: manzyuk

Differential Revision: D13449702

fbshipit-source-id: 9152c6150211abf2a94f11896cda33da14b96e6f