disjunction of constraints: take the first one that works
Summary:
This is better than the current state of the new inference algorithm, where we completely ignore any proposition that's in a disjunction. This is not worse than the legacy, which would also take the first thing that works. Actually, it's more complete than the legacy on the added test cases, where the legacy would have some false positives.

There is obviously room for improvement on completeness, as we may discard propositions too early, but it's at least as good as the legacy, so we should save improvement on completeness for later.

Reviewed By: andrewjkennedy

Differential Revision: D13527998

fbshipit-source-id: 078da3fce5ad0060ac893504e5d0767f96813bea