Only emit half the body for Rx\IS_ENABLED __Rx functions/methods.
Summary:
Detect at emission time if we're emitting an Rx function/method whose toplevel is a single if/else statement, where the if is testing exactly the HH\Rx\IS_ENABLED constant (with proper namespacing support), and the else is not empty. Having proper namespacing support requires auto-importing the Rx\IS_ENABLED constant, the first constant to be auto-imported from the HH namespace in this way.

If we detect an eligible function/method, then use a new RxIsEnabled runtime option to emit the appropriate half of the body. Since it affects emitter behavior, hash this option into the unit hash.

If RxIsEnabled was false and we emitted the non-Rx half of a function using if (HH\Rx\IS_ENABLED), then leave the Rx attributes of the function unchanged, but tag it with an isRxDisabled flag. Propagate the isRxDisabled flag through hhas (and hhbbc) into the runtime and use it to disable Rx body checks in the verifier. The flag is also available for use in disabling calling convention enforcement for calls made from inside the body of an isRxDisabled function, but that calling convention work was not done on this diff.

Delete the ext_rx extension that was providing the HH\Rx\IS_ENABLED constant to the runtime based on the value of the RxPretendIsEnabled option. The constant is no longer available at runtime, but the RxPretendIsEnabled option still exists for migration purposes, and the value of the new RxIsEnabled option defaults to the value of the old one.

Reviewed By: oulgen

Differential Revision: D13525464

fbshipit-source-id: a747767cd1c7c4ff0ff841b4c8ad73f45c50a390