Fix braindead bug in array_slice
Summary:
Saw this when looking at a perf regression in array_slice--200% not
sure what I was thinking here but this is obviously wrong. I went back and
reviewed D13110799 (where this was added) to see if I had done anything else
braindead like this and didn't see anything.

This is an optimization in array_slice that rebuilds the array starting
0-indexed if the user allows it (according to the preserve_keys argument).
Because of this mistake, we would fail to preserve the keys if the caller
asked, provided a packed array, did not request the whole collection back, and
EnableIntishCast was false.

Reviewed By: markw65

Differential Revision: D13528174

fbshipit-source-id: 5b29f7961cde63cdfa74ae543fb82ed0da651c24