Fix AbortBuildOnVerifyError when --hphp -vUseHHBBC=false
Summary:
Add show helper for repo global data, and use it to print the global data we loaded from the repo when TRACE=hhbc:1

This showed that we weren't copying AbortBuildOnVerifyError into the repo global data in hphp/compiler/analysis/emitter.cpp, so when we ran tests with --repo (which uses --hphp -vUseHHBBC=false to compile), the repo would always have AbortBuildOnVerifyError = false in the global data, and we would never run the verifier when running --hhbbc in step 2.

By contrast, in --repo-single mode we didn't rely on the value stored in the repo when running --hhbbc, so we would run the verifier and could get hhbbc aborts when the test failed validation.

Fix it by correctly copying AbortBuildOnVerifyError into the repo global data in hphp/compiler/analysis/emitter.cpp.

This fix revealed another problem: the code that tries to verify units after they come out of hhbbc when AbortBuildOnVerifyError is on will always assert, because verification relies on being able to read from the LitstrTable, but committing units to the repo requires being able to write to the LitstrTable, and we currently do not allow both reads and writes from/to the LitstrTable at the same time. We were getting away with this because the verification code was missing from the code path that commits units when we compile and optimize in one step, and the bug that this diff is fixing meant the code trying to do this in hhbbc/main.cpp (when optimizing as a separate step) was never hit. (There was a further problem with the code in hhbbc/main.cpp: it was only run outside the main unit commit loop, so it would only try to verify the last 1-7 units.)

Because of this problem with verifying units while committing them, remove the code in hhbbc/main.cpp that tries to do so. We will need to figure out a way to safely read litstrs for verification purposes while committing units to the repo, and then we can re-introduce the verification code (hopefully in a code path that is shared between 1-step and 2-step compile/optimize, so that we can't accidentally leave it out of one of them again). I've made T38383456 for this follow-up work.

Reviewed By: markw65

Differential Revision: D13535506

fbshipit-source-id: 25a52f4323fcd719b69154142f725b65ce53fa27