Improve SetOp performance when CheckPropTypeHints is enabled
Summary:
When performing a SetOp operation when CheckPropTypeHints is enabled, we might
have to perform the SetOp on a temporary location, do the type-hint check on
that, and if it passes, move the temporary to the final location. This is
because if the type-hint fails, we'll re-enter to raise a warning, and it might
be possible to access the property (and it cannot ever contain a value of the
incorrect type). This is fine for most operations, but its bad for concats. This
is because the temporary increases the ref-count of the value, which means the
concat will now trigger a COW. This is especially bad for workloads which
repeatedly append to a string, as every append will now incur a copy.

Fix this by checking if the concat is safe to perform without a type-hint check
before performing the SetOp. If the lhs of the concat is already a string or if
the type-hint always allows strings (concats always produce strings), then the
type-hint check will always pass and we don't need to use the temporary. We can
add similar skips to other SetOps as necessary.

This optimization speeds up a synthetic benchmark by ~50,000x with
CheckPropTypeHints enabled.

Reviewed By: markw65

Differential Revision: D13540524

fbshipit-source-id: 08b673637c7eca067724fb9c84e1bc3b1d263358