Move client functions from `ServerCommand` to `ServerCommandLwt`
Summary:
The client uses Lwt, so the synchronous `wait_for_rpc_response` is no longer called. We may as well remove it, and then move the rest of the functions to `ServerCommandLwt`, where the rest of the client API lives.

We could also update the rest of the functions called by the client to Lwt, but I think that's not necessary at this point. This could be a confusing name for the module, though; I'm open to other suggestions. (`serverClientCommand`? Also sounds confusing.)

Reviewed By: 2BitSalute

Differential Revision: D13440937

fbshipit-source-id: bb5d601d09507292c17a2d02f42cd25537859a05