Fix SetElem HHIR return type
Summary:
D9123613 changed SetElemString to always return a static string instead of a
counted string. While this is probably the better behavior, the JIT specifically
assumed a counted string would always be returned. As a result of this, we can
now return a static string, which the JIT assumes is counted, meaning it will
perform ref-count operations on it, leading to static strings with garbage
ref-counts. I'm surprised this hasn't caused issues until now, but the
conditions under which we actually return a string are rare. However, with
property type-hint enforcement on, it causes problems immediately. Change the
JIT to match the new runtime behavior.

Reviewed By: markw65

Differential Revision: D13566920

fbshipit-source-id: 01185e5d83f40b19bee7589139c6e3d52161be6b