Fix crash when using DateTime instances without running ctor
Summary:
DateTime, DateTimeZone, and DateInterval all initialize their native data inside
their constructors (not their native instance constructors). This is apparently
because initializing the data depends on parameters passed to the
constructor. However, like any PHP object, the constructor is not guaranteed to
have been called when the object is used. As a result, if the constructor is not
called, the objects will not be initialized properly and can segfault when
trying to use them. This can be easily triggered by subclassing from DateTime,
DateTimeZone, or DateInterval and neglecting to call parent::__construct().

Fix this by having the various methods which use the native data check if the
data is initialized (this is easy because its stored using a unique_ptr). If it
is not, throw an exception (as opposed to just crashing).

Reviewed By: markw65, mofarrell

Differential Revision: D13568106

fbshipit-source-id: 521e7c0a30305ded43c90aa1c9c2de0672ed5332