Fix assertion when throwing from intercept handler
Summary:
When we re-enter the VM, we record the current stack pointer location. When we
return, we'll assert in debug builds that the stack pointer matches what it was
before we re-entered. Importantly, the stack pointer is recorded before we push
any placeholder in-out stack slots. If we return normally, we'll pop these slots
off. However, if we throw from the re-entry, we will not pop them (and the
unwinder won't either). This means we'll trigger the assert as the stack pointer
won't match at the point of the check. This can be easily triggered from
throwing from inside the intercept handler (which both takes in-out parameters
and which forces a re-entry).

Fix this by using SCOPE_FAIL to pop off the in-out placeholder slots in the case
of an exception.

Reviewed By: markw65

Differential Revision: D13567596

fbshipit-source-id: 294373cf5ccc9c4d3b476ae5a6897e31584b6fc1