Clarify error message for PHP7 anonymous functions
Summary:
Show a clearer error message when a user puts the type annotation in the wrong place.

```
// Correct
$f = function(): void use() {};

// Syntax error
$x = function() use(): void {};
```

The latter is valid PHP 7 syntax, so we would report "PHP 7 anonymous functions are not supported". This is easily confused with "PHP anonymous functions are not supported".

Change the error to be simpler, and more actionable.

Reviewed By: dlreeves

Differential Revision: D13558283

fbshipit-source-id: 1d1727311cb31ca9d79c6b4919259e63a3fc4f62