Change constructor of Aast.CI to only take Aast.sid
Summary:
D13340822 introduces a `targ list` on the `Aast.New` constructor, however `Aast.New` also takes a `class_id` where the `Aast.CI`:

```
CI of sid * targ list
```
This creates a situation where `Aast.New` can be passed a `targ list` via the `class_id` or as an argument to the constructor.

Looking at usages of `CI`, we only populate the `targ list` when needing a `class_id` for `Aast.New`, otherwise we mostly ignore it.

Consequently, this diff updates the `Aast.CI` constructor to:

```
CI of sid
```

Most of the code stays the same, but for `Aast.New`, we pass the `targ list` as an argument to the constructor instead of via the `class_id`. E.g.

```
Aast.New (CI (sid, tal), [], ...)

becomes

Aast.New (CI sid, tal, ...)
```

Reviewed By: Wilfred

Differential Revision: D13565076

fbshipit-source-id: f2b4ea9648b7c7e6be8d198c9e4d81ed8576d4ad