Kill stack/local equivalencies smarter in HHBBC
Summary:
I was noticing a lot of PopL; PushL pairs (with the same local) when looking at
bytecode. HHBBC does eventually remove these pairs during DCE, but before that
they pessimize analysis a bit. Namely, if we have a (simplified) sequence like:
CGetL $a; PopL $b; PushL $b by the time we're done we'll have forgotten that the
top of the stack is equivalent to $a even though it is. All is well until the
PushL. At that point we know $a and $b are equivalent. The PushL, however is
modelled as CGetL (top of the stack is equivalent to $b), and UnsetL (which
kills any equivalences with $b). This just sets the equivalency at the top of
the stack to NoLocalId, destroying the fact that the top of the stack was also
equivalent to $a.

Fixing this is simple. When killing stack equivalencies, don't unconditionally
set it to NoLocalId. Instead check if there's another equivalent local, and set
it to that instead. We also need to kill stack equivalencies before killing
local equivalencies. While we're at it, reduce UnsetL to a Nop if the local is
already uninit (which it usually is).

Reviewed By: markw65

Differential Revision: D13570095

fbshipit-source-id: fab77f5a6cea1e7120762d638257dd15b20c4720