Field initialization with case statements
Summary:
Rough draft for solving the bug with initialized properties not working in case statements with cases that fallthrough.

There's a bug with the typechecker, that when a switch statement has a case that falls through, it will mark a field as not being set. Examples:

```
// this is okay
final class B {
  protected int $batchID;

  public function __construct(string $batch_id) {
    switch ($batch_id) {
      case 'test2':
        $this->batchID = 1;
        break;
      case 'test3':
        $this->batchID = 2;
        break;
      default:
        $this->batchID  = 3;
    }
  }
}

// this is an error where batchID isn't set

final class B {
  protected int $batchID;

  public function __construct(string $batch_id) {
    switch ($batch_id) {
      case 'test2':
      case 'test3':
        $this->batchID = 2;
        break;
      default:
        $this->batchID  = 3;
    }
  }
}
```
Related: https://our.intern.facebook.com/intern/diff/D13407017/

Reviewed By: kmeht

Differential Revision: D13523571

fbshipit-source-id: 283e2570260719cf4dc74a6974203a5aa0de76af