Fix some idx issues in the gdb helpers
Summary:
We used a regex in idx to quote the "main" part of the type; but it
didn't work if there was a pointer type in a template param, and it
didn't work for const pointers.

eg

  std::__atomic_base<HPHP::Func *>::__pointer_type *

would be quoted as

  'std::__atomic_base<HPHP::Func const '*>::__pointer_type *

rather than

  'std::__atomic_base<HPHP::Func const *>::__pointer_type' *

Similarly `HPHP::Func const *` would be quoted as `'HPHP::Func const' *`
rather than `'HPHP::Func' const *`

In addition, if the current language was set to C rather than C++,
template_type would return "struct Foo" or "class Foo" rather than
"Foo", breaking a number of direct string comparisons we use to
identify types.

Reviewed By: mofarrell

Differential Revision: D13573051

fbshipit-source-id: 9d4fdb4191e0824b39c8928a09b59164d5315085