Deal with magic interfaces properly when property type checking is enabled
Summary:
When the JIT is calculating an invariant type for a property based on its
type-constraint, it was neglecting to deal with magic interfaces. As a result,
something like Traversable would be considered an object, when it could be an
array, for example. If we infer a magic interface, just treat it as TInitCell.

Reviewed By: paulbiss

Differential Revision: D13573645

fbshipit-source-id: f6bf9f31d879a9ce04a4b2c0acbe5d10a33e5c93