Deal with mixed property type-hints properly in JIT
Summary:
A type-hint to mixed does not forbid binding a property to a ref, so the
inferred type in the JIT should be TGen. Moreover, if we can't adequately
resolve a type alias, we have to assume it might be mixed so likewise treat it
as TGen.

When we assert on the type within a property, we need to do a conservative
"maybe mixed" check and allow a ref if it returns true.

Reviewed By: paulbiss

Differential Revision: D13573933

fbshipit-source-id: 6c899159fbfea68b0919bc3d109d88fefbfd7e10