Fix HHBBC inference of whether a type-constraint implies InitCell or Cell
Summary:
HHBBC was assuming that only type-constraints which can be resolved to a mixed
annot-type could be mixed. If the type-constraint could not be mixed, it implies
the admissable type is TInitCell (since any type-constraint except mixed forbids
uninit). Of course, just because a type-constraint cannot be resolved doesn't
mean it can't still be mixed.

Fix by this distinguishing the different cases for returning folly::none in
get_type_for_annotated_type(). Namely, we want to distinguish returning
folly::none because we cannot express the admissable type in HHBBC's
type-system, and folly::none because we cannot resolve the
type-constraint. Return a boolean which indicates that the type-constraint might
possibly be null, and use that to either resolve to TInitCell or TCell.

Reviewed By: markw65

Differential Revision: D13568602

fbshipit-source-id: 20f9f623697d4d17aa0327029b2b20c7afbbde44