RDS local buffer cleaned up at matching call
Summary:
If the rds local area is allocated through a call to `rds::local::init`, it
should be cleaned up by the matching call to `rds::local::fini`.  Similarly
so for `hphp_thread_init`, and `hphp_thread_exit`.  This will make it so the
lifetime of the section is not cut short unexpectedly.

Reviewed By: binliu19, markw65

Differential Revision: D13528798

fbshipit-source-id: 6c11599277adeb96e2c791d0d7d7f465e3d119ba