Mark require-extends and require-implements ancestors as synthesized in linearization
Summary: In some situations (see added test case), it is necessary to distinguish between ancestors which were included because of a require-extends or require-implements relationship and ancestors which weren't. This change marks ancestors included because of a require-extends or require-implements relationship as synthesized in the linearization, allowing consumers to filter them out when appropriate.

Reviewed By: vassilmladenov

Differential Revision: D13409583

fbshipit-source-id: 0a2223ddfbd1878817274264711a5ee328df7860