Tobject should not be a subtype of Tclass in strict mode
Summary:
The following example is currently accepted by the type-checker:
```
<?hh // strict
// Copyright 2004-present Facebook. All Rights Reserved.

class C {}

function test<T>(classname<T> $c): C {
  return new $c();
}
```
We infer the type `Tobject` for `new $c()` and `Tobject` is considered to be a subtype of `Tclass _`, even in strict mode.  (It goes like this: in `simplify_subtype` we say that `Tobject <: Tclass _` is valid in non-strict modes and in strict mode we default to `sub_type_inner_helper`, which in this case tries to unify `Tobject` and `Tclass _`, and the unification succeeds.)  This example is clearly unsound and should be rejected.

Note that if `T` is constrained by a class or an interface, we infer a dependent type for `new $c()`, so this unsoundness doesn't arise.

Reviewed By: andrewjkennedy

Differential Revision: D13528069

fbshipit-source-id: 6baface18aeb3f44feabea48255625ef12ff94a8