Correctly set the variance of type variables
Summary:
The way we set the variance of type variables is not entirely correct at the moment.  For example, consider the following code:
```
<?hh // strict
// Copyright 2004-present Facebook. All Rights Reserved.

class C<T> {}
class Inv<T> {}

function f<T>(Inv<C<T>> $_): void {}

function test(): Inv<C<int>> {
  $x = new Inv();
  f($x);
  return $x;
}
```
The new inference algorithm rejects this code.  It's instructive to work out why:
- We instantiate the type parameter `T` in the definition of `Inv` with a fresh type variable `v`, that `$x : Inv<v>`.
- We instantiate the type parameter `T` in the definition of `f` with a fresh type variable `w`, and the call `f($x)` gives rise to a constraint `Inv<v> <: Inv<C<w>>`, which simplifies to `C<w> <: v && v <: C<w>`.
- Because `w` doesn't appear in the type of the expression `f($x)`, we solve for `w` by pushing down to the bottom type.
- Now we have `C<_|_> <: v && v <: C<_|_>` and the return statement gives rise to a constraint `v <: C<int>`, which by transitivity implies `C<_|_> <: C<int>`, which simplifies to `_|_ <: int && int <: _|_`.  The second constraint in this proposition is obviously unsatisfiable.

The problem here, it seems, is that despite having a constraint `v <: C<w>` where `v` is invariant, we still treat `w` as covariant and solve for `w` too early.

In [[ https://ac.els-cdn.com/S0890540101929633/1-s2.0-S0890540101929633-main.pdf?_tid=c6beca95-56d1-4b70-916e-2460a3c32582&acdnat=1543336531_f66d9412943f58ae6f463fce3d87980f | "Simplifying Subtyping Constraints" ]], Pottier defines (in Definition 4.4) polarities of type variables in a set of constraint in such a way that, for example, a type variable occurring covariantly in an upper bound of type variable variable marked as contravariant must itself be marked as contravariant.  This diff is trying to replicate Pottier's definition.

In his setup, polarities are constructed once, and this can be done in linear time in the size of the constraint set: we need to visit a lower (resp. upper) bound of a variable only once, namely when the variable first is marked as covariant (resp. contravariant).  Because we try to solve constraints as we go, we need to recompute the polarities as new variables are added to the type variable environment and new constraints are added to the subtyping proposition.  In the above example, the variable `v` is added to the type variable environment and is marked as invariant when we type-check the expression `new Inv()`.  We have to revisit `v` when `w` is introduced because `v` now has new bounds and invariance of `v` needs to be transferred to the variables occurring in those bounds.

Reviewed By: andrewjkennedy

Differential Revision: D13466536

fbshipit-source-id: 9accaa7bf7691865489221ce0c3f992a0523141a