New inference: Revert unsound array append change
Summary:
An earlier attempt to implement constraint-based array append (D13449702) was actually unsound. For copy-on-write arrays, it's not sound to use a signature such as
```
append<Ta as toparray<Tv>, Tv>(Ta $array, Tv $value): Ta
```
For example, consider instantiating `Ta` with `vec<int>` and `Tv` with `arraykey`. The return type of this should be `vec<arraykey>`, not `vec<int>`!

It would be sound for mutable collections to use a signature
```
append<Ta as toparray<Tv>, Tv>(Ta $array, Tv $value): void
```
but the whole point of the original diff was to handle *all* overloads of append with a single signature. In the absence of intersection types on functions, I can't see how to do this.

For now, let's revert to "eager" solving of arrays that have type variable type. (I'm leaving all the `tyvars` plumbing in place, in case we need to use this later.) To this end, I've moved a couple of tests from the original diff into the `eager_solve` subdirectory that sets the appropriate flag.

There remain some examples that get rejected (see `ref_get_array_append.php` in the original diff). The problem there is that we have a lower bound such as
```
vec<nothing> <: #1
```
which gets solved eagerly to the highly-specific `#1 := vec<nothing>` in order to type-check an append operation. But observe that append only needs the "head" type constructor in order to make progress: in this case, `vec`. So we could instead solve eagerly with `#1 := vec<#2>` for a fresh type variable `#2`. Further constraints would be imposed on `#2` through other subtype checks. I will work on this idea in a separate diff.

Reviewed By: CatherineGasnier

Differential Revision: D13579947

fbshipit-source-id: 034b50c5a6b7f93113eb36a48c0b3cb89ed0fd52