New Inference: Don't check type of index against collection key type for covariant keys
Summary:
Legacy arrays and Hack arrays are covariant in their key types. In an expression `e1[e2]`, it's brittle to reject the index if the inferred key type of `e1` is not a supertype of the inferred index type `e2`. Why? Because by covariance, the key type of `e1` can always be pushed to the "top" type for the particular collection, e.g. `arraykey` for `dict`s, and it's always legal for type inference to infer such a key type. So for example, if Hack infers `dict<int,bool>` for `e1`, and infers `string` for `e2`, it should not reject the program, because a valid typing of this code assigns `dict<arraykey,bool>` to `e1`.

In this diff we disable this checking for copy-on-write arrays that have covariant key types, and simply check that the index expression has a type that is a subtype of the "top" type for the collection. We do this if `new-inference` is set, because otherwise many false positives result from the constraint solver producing a very "tight" type for the collection.

Note that in principle we should actually do the same for reference-semantics collections such as `Map`, because these implement an interface (`Indexish`) that is itself covariant in the key type, and we can legally use this interface to index the collection. But for the sake of legacy behaviour, and the absence of false positives appearing in practice, we do not yet disable the index check for reference-semantics collections.

Note: in the longer term, it makes much more sense to make the key type *invariant* for all arrays and collections. However, this is likely to produce a lot of new errors, so we postpone this change for now.

Reviewed By: manzyuk

Differential Revision: D13590142

fbshipit-source-id: 2596485c0b2ec78e26527dbc947d029f40806ce9