Improve error message for static-dynamic member name conflicts
Summary:
These error messages are a little unclear. Consider this example:

```
class ABC {
  protected static function staticMethod(): void {}
}

class DEF extends ABC {
  protected function staticMethod(): void {}
}
```

Previously, Hack reported an error on ABC::staticMethod, saying, "the method staticMethod is static; it is defined as non-static here: DEF::staticMethod". Since the error is reported in ABC, a user investigating the error will be taken to ABC, and might reasonably wonder why DEF's definitions are any of ABC's business.

This happens because we emit the error when typechecking DEF, but we emit it at the position of the static method, wherever it is (even if ABC is in another file). Because we always report the error at the position of the static method, in cases like this, we seem to be reporting that the declaration of ABC::staticMethod is in error, when in fact DEF's definition is the erroneous one, since it is attempting to replace an inherited static method with an instance method.

After this change, the error is always reported on the child class, is only detected when typechecking the file in which the child class is declared, and is reworded to be a little clearer.

Reviewed By: Wilfred

Differential Revision: D13586277

fbshipit-source-id: 701fd58a554ebefa0e1f816eff60bf4fb46fc973