New inference: less constraining eager solving
Summary:
The "eager" solving of type variables, used for type-checking overloaded operations such as array append and update, and function application, was *too* eager.

For example, suppose that we have a lower bound such as
```
vec<int> <: #1
```
which gets solved eagerly to `#1 := vec<int>` in order to type-check an append operation. This unfortunately results in the rejection of some programs, where the solution should be a wider type such as `vec<arraykey>`.

Observe that overloaded operations only need to force the "head" type constructor, here `vec`. (For function application, we only need know that it is a function type of some kind.)

So instead, we can solve eagerly with `#1 := vec<#2>` for a fresh type variable `#2`, with the additional constraint that `int <: #2`. Further constraints would be imposed on `#2` through other subtype checks, for example, resulting in the solution `#2 := arraykey`.

To implement this, we define a function `freshen_inside_ty_wrt_variance` that replaces all variant "components" of a type with fresh type variables, and applies appropriate subtype assertions on these fresh variables. Note that unions and nullables are treated specially, with the freshening happening underneath the union. (For example, `?vec<int>` gets freshened to `?vec<#1>` and not to `?#1`.

Reviewed By: manzyuk

Differential Revision: D13580330

fbshipit-source-id: 01f264cb00b677e9b0e3fc30fc972dda676e3a70