Eagerly solve the type of the RHS of a list-destructuring assignment
Summary:
At the moment, when we type-check an assignment `list(e1, ..., en) = e;`, we do case analysis on the type `t` of `e` in order to extract the types of `ei`, e.g., if `e` is an array or `Pair`; and in the default case, we create a tuple type `(v1, ..., vn)` with fresh type variables `vi` and assert that `t` is a subtype of `(v1, ..., vn)`.  This approach is known to have limitations even in the legacy type-checker (e.g., we can't destructure values of unresolved types, see T28515957), and in the new inference algorithm we run into a problem when `t` is a type variable.

I'm not sure yet what the right approach to type-checking such assignments is, but in short term we could expand and solve the type `t`.

Reviewed By: andrewjkennedy

Differential Revision: D13608383

fbshipit-source-id: d89c148f9d2d2140d96c410ce6f8b5004ccf377e