Lower and name file attributes
Summary:
This diff lowers and names file attributes. Here's how it works:

In lowering, just produce an `Ast.FileAttributes` for every file attribute node in the CST. Leave `f_file_attributes` and `c_file_attributes` empty for the Ast functions and classes.

In the early elaboration step `Namespaces.elaborate_top_level_defs`, fill in the `fa_namespace` field of the Ast node, but do not elaborate the names of the actual attributes in `fa_user_attributes`.

Still in `Namespaces`, pass over the list of toplevel declarations and collect all of the `Ast.FileAttributes` nodes.

Pass over the toplevel declarations again, copying all of the file attributes into `f_file_attributes` and `c_file_attributes` on the Ast nodes for classes and functions.

In naming.ml, use the Ast node's `fa_namespace` to properly name the `fa_user_attributes`. At this point, all of the names in the user attributes have been fully elaborated, so we don't need to keep the `fa_namespace` information separately. We end up with just a `user_attribute list` field in Aast nodes for classes and functions.

Differential Revision: D13068639

fbshipit-source-id: afeef8c453f441bfc61cea295324cd80eac30440