Constrain MBase guards properly
Summary:
We weren't constraining guards on the MBase properly. It worked if you
constrained the guard before the first usage of LdMem on a pointer to the MBase,
but not afterwards. This is because the LdMem becomes the "type-source" for the
MBase and we chain the constriction through that instruction. However the guard
constraint logic doesn't deal with LdMem at all, so the constriction goes
nowhere. The solution is to have the guard constriction logic deal with LdMem in
a special way. If the LdMem is using a pointer derived from LdMBase, it should
be treated as a load from the MBase in exactly the same way as LdLoc or
LdStk. That is, constraining through the LdMem should chain back to the MBase
and constrain there.

This was noticed while enabling hard prop type-hint enforcement. The minstr code
was trying to constrain the object type in the MBase (because it was using the
type information gleaned from it), but the constraint wasn't taking and the
object type was being relaxed away, which then caused incorrect code to be
generated. Also IncDec was attempting to only do a weak constraint, but we
actually want a normal constraint here. With that, there's no need for the
"constrain" parameter, so we can remove it.

Reviewed By: markw65

Differential Revision: D13599525

fbshipit-source-id: 19ab645769cd164eb3e4b7b8955a5b739a4503f7