fix stack overflow
Summary:
Say you have
```
```

Then you add a nonnull lower bound A to #1.

It's not already a lower bound of #1, so you go ahead an add it.
Since you do `try_union` on lower bounds of #1, you don't actually add it because it is a subtype of nonnull.
Then you do the transitive closure: you add A to lower bounds of #2.

It's not already a lower bound of #2, so you go ahead an add it.
Since you do `try_union` on lower bounds of #2, you don't actually add it because it is a subtype of nonnull.
Then you do the transitive closure: you add A to lower bounds of #1.

It's not already a lower bound of #1, so you go ahead an add it.
...

stack_overflow.

This diff fixes that by generalizing the "It's not already a lower bound of #_" part of the algorithm: see first Typing_subtype.add_tyvar_upper_bound (the crux of this diff).

Reviewed By: manzyuk

Differential Revision: D13542458

fbshipit-source-id: 3c66a6cf131753aa735ccb4ef0ac2d384ec4706c