Check arity for all generics
Summary:
Hack was only checking the top level of generic arity. In practice, this meant that

```
class C<T> {}
...
new C<C<int, string>>();
```
was not an error. This change forces that with erased generics, they are either all specified or not present.

With reified generics, they are forced to always be present.

Reviewed By: dlreeves

Differential Revision: D13611505

fbshipit-source-id: 79fb289fc827ecbfe09518687e4c70de50d692f5