Use O(1) context in FFP errors' has_inout_params & is_inside_async_method
Summary:
In light of shifting error logic to "smart" constructors,
we want avoid linear-time checks on AST that go up (e.g., parents)
and eventually replace it by context/visitor that passes it down.

This diff replaces lookup in variable `parents` with `context.active_callable`,
which is guaranteed to take O(1) time and works with smart constructors.

Reviewed By: oulgen, Wilfred

Differential Revision: D12934494

fbshipit-source-id: da06b5d41f7ffe749cb0d0abdc61d36148132886